finished ver1
